Implementation of data structure verification
Included is the complete implementation of both the one-dimensional
list and Merkle tree data structures (with both single leaf and all-leaf
verification for the Merkle tree structure).

The result of gas testing shows the following figures, for a list of 8
items:
i. verifyMessageArray (all items verified): `28505`
ii. verifyMerkleAll: `29728`
iii. verifyMerkleLeaf: `25423`